backend/requirements.txt
annotated-types==0.7.0
bcrypt==4.0.1
blinker==1.9.0
cffi==2.0.0
click==8.3.2
cryptography==46.0.7
dnspython==2.8.0
dotenv==0.9.9
ecdsa==0.19.2
email-validator==2.3.0
Flask==3.1.3
greenlet==3.4.0
gunicorn==25.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
MarkupSafe==3.0.3
marshmallow==4.3.0
packaging==26.0
passlib==1.7.4
psycopg2==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
python-dotenv==1.2.2
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.8
